+++ src/Api/Strategies/Strategies.ts
@@ -1,6 +1,7 @@
 
 import { Strategy } from "../Types";
 import checkForSolved from "./checkForSolved";
+import hiddenPairsTriplesAndQuads from "./hiddenPairsTriplesAndQuads";
 import hiddenSingles from "./hiddenSingles";
 import pairsTriplesAndQuads from "./pairsTriplesAndQuads";
 import updateCandidates from "./updateCandidates";
@@ -11,6 +12,7 @@ const STRATEGIES = [
    updateCandidates,
    hiddenSingles,
    pairsTriplesAndQuads,
+   hiddenPairsTriplesAndQuads
 ] as const
 
 export default STRATEGIES as typeof STRATEGIES & Strategy[]